requirements.txt
flask
flask-cors
requests
python-dotenv
time